Initial commit: make first pass at server, build basics
Implement a first pass at a CSP-style, UNIX-domain server which
mimics an old-style, 033-taught Ethernet link. Given a message from
a given client, simply repeats it to every other client, including
the client who initially sent the message. This allows the actual
client implementation to remain as true to the Ethernet spec as possible,
and realistically emulate what we're trying to get across.

Also add some build files, a license, README, etc., along with the basics
for command line configuration.

Signed-off-by: Jay Lang <jaytlang@mit.edu>