Write the tap side and complete the basic setup
Write a very simple, work-in-progress TUI which allows a user
to connect to the server we wrote last commit, and send messages.
There's a bit of an odd (but obvious) interleaving when the server
sends back the same output as we do, and there's no carrier sense
implemented yet. These things are logical next steps and shouldn't
be terribly hard.

Signed-off-by: Jay Lang <jaytlang@mit.edu>