Prepare for CRS by passing *net.Conn to broadcast
The broadcast method now doesn't send messages back to a server
which originally sent them. This is far superior to the original
model in terms of ease of use while still staying relatively true
to the ethernet model.

The rest is pretty straight forward. We should be able to just hook
up a channel between the tap's send/recv and issue a kill signal if
we see problems.

Signed-off-by: Jay Lang <jaytlang@mit.edu>