Implement collision detection
Observe that if we monitor the output from sending and receiving
status channels, which each output `true` when new data is coming over
the line and `false` when it completes, that we can detect collisions
by watching for two `true`s in a row.

Using this insight, we can implement basic collision detection. From here,
all that remains is adding sort of forced sleep to the in and out threads
to ensure they shut up while exponential backoff takes place.

Signed-off-by: Jay Lang <jaytlang@mit.edu>