Implement line jamming
Jam the line with a skull-and-crossbones guy to force collision
consensus enforcement, in accordance with 3.5.5 of the paper.
This has some interesting side effects that dont' work all the time,
especially due to the weak assumption we use to detect collisions,
but for the purposes of a monday demo it works.

This completes the initial prototype.

Signed-off-by: Jay Lang <jaytlang@mit.edu>