Add message corruption
Adding a backtick (`) before a given character in a message will
corrupt it. This forces the CRC32 checksum to fail. Implement
this functionality...sort of.

In actuality, make the tap intercept backticks and append a "✓"
character if none are detected (as before), or otherwise an 'X'
character. Have the receiver then determine which message terminator
is used and signal the user accordingly.

Signed-off-by: Jay Lang <jaytlang@mit.edu>