Add visible message corruption for receivers of backtick messages
Before, corruption just triggered a "CRC32 failure", even though
they received the intact message. This makes no sense, so add
corruption of messages by adding/subtracting random offsets off
the rune before encoding it for the wire, but after printing it.

The corruption hits a lot closer to home as a result, so should
be easier to understand.

Signed-off-by: Jay Lang <jaytlang@mit.edu>